+++ service/ControlSituationAdmin/src/main/java/com/control/situation/entity/UserRoleInfo.java
@@ -21,41 +21,40 @@ public class UserRoleInfo implements Serializable {
 	// 主键
 	@Id
 	@Column(name = "id")
-	private Integer id;
+	private Long id;
 
 	// 用户ID
 	@Column(name = "user_id")
-	private Integer userId;
+	private Long userId;
 
 	// 角色ID
 	@Column(name = "role_id")
-	private Integer roleId;
+	private Long roleId;
 
 
 	public UserRoleInfo(){}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
 
-	public Integer getRoleId() {
+	public Long getRoleId() {
 		return roleId;
 	}
 
-	public void setRoleId(Integer roleId) {
+	public void setRoleId(Long roleId) {
 		this.roleId = roleId;
 	}
-
 }
